+++ internal/config.go
@@ -7,7 +7,7 @@ const (
 
 // 默认值
 const (
-	DefaultVersion         = "v0.1.2"  // 版本号 默认值
+	DefaultVersion         = "v0.3.0"  // 版本号 默认值
 	DefaultPort            = 80        // 端口 默认值
 	DefaultTargetPort      = 80        // 目标端口，默认值
 	DefaultDataPath        = "./data/" // 数据存储目录路径 默认值
@@ -36,6 +36,7 @@ const (
 const (
 	CodeOk  = 200
 	Code200 = CodeOk
+	Code400 = 400
 	Code404 = 404
 	Code500 = 500
 )